+++ src/components/interaction-designer/clipboard/prompt-kinds/Numeric.vue
@@ -37,34 +37,17 @@
 import BlockActionButtons from '../shared/BlockActionButtons.vue'
 import Component, { mixins } from 'vue-class-component';
 import Lang from '@/lib/filters/lang';
-import { Prop } from 'vue-property-decorator';
-import { IPrompt } from '@floip/flow-runner';
-import { namespace } from 'vuex-class';
-
-const clipboardVuexNamespace = namespace('clipboard')
+import { PromptKindMixin } from '@/components/interaction-designer/clipboard/shared/PromptKindMixin';
 
 @Component({
   components: {
     BlockActionButtons
   },
 })
-export default class Numeric extends mixins(Lang) {
-  @Prop() index!: number
-  @Prop() isComplete!: boolean
-  @Prop() goNext!: Function
-  @Prop() onEditComplete!: Function
-
-  isBlockInteraction = false
+export default class Numeric extends mixins(Lang, PromptKindMixin) {
   enteredValue = ''
   backUpValue = ''
   errorMsg: string | null = null
-
-  get isFocused() {
-    return this.isBlockFocused(this.index)
-  }
-  get prompt() {
-    return this.getBlockPrompt(this.index)
-  }
 
   checkIsValid() {
     const num = +this.enteredValue
@@ -103,11 +86,5 @@ export default class Numeric extends mixins(Lang) {
     this.enteredValue = this.backUpValue
     this.errorMsg = ''
   }
-
-  @clipboardVuexNamespace.Getter isBlockFocused!: (index: number) => boolean
-  @clipboardVuexNamespace.Getter getBlockPrompt!: (index: number) => IPrompt<any>
-  @clipboardVuexNamespace.Action setIsFocused!: (data: { index: number, value: boolean }) => void
-  @clipboardVuexNamespace.Action setLastBlockUnEditable!: () => void
-  @clipboardVuexNamespace.Action setLastBlockEditable!: () => void
 }
 </script>
